remove a staled socket file on start

+++ spec/backend/unix_socket_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe BJob::Backend::UNIXSocket do
   subject { described_class.new(path: socket_path, coordinator: coordinator) }
 
   after do
-    FileUtils.rm(socket_path)
+    FileUtils.rm(socket_path) if File.exist?(socket_path)
   end
 
   context 'with valid JSON formatted request from client' do
@@ -25,6 +25,16 @@ RSpec.describe BJob::Backend::UNIXSocket do
 
     def send_request
       UNIXSocket.new(socket_path).puts(client_request)
+    end
+  end
+
+  context 'socket file already exists in the system' do
+    before do
+      FileUtils.touch(socket_path)
+    end
+
+    it 'removes socket file if it exists' do
+      expect{subject.start}.not_to raise_error(Errno::EADDRINUSE)
     end
   end
 end
